# swagger/src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"CountResponse": {
				"description": "Response for general count",
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DeleteResponse": {
				"description": "Response for how many records have been deleted.",
				"properties": {
					"deletedCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"deletedCount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EntityOverallData": {
				"description": "It stores id and name. Used to minimize entity.",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RouteResponse": {
				"description": "response of route apis",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"siteId": {
						"type": "number",
						"format": "double"
					},
					"allowStartTime": {
						"type": "number",
						"format": "double"
					},
					"patrolMode": {
						"type": "string"
					},
					"site": {
						"$ref": "#/components/schemas/EntityOverallData"
					},
					"createdUserId": {
						"type": "string"
					},
					"createdUserName": {
						"type": "string"
					},
					"assignedUsers": {
						"items": {
							"$ref": "#/components/schemas/EntityOverallData"
						},
						"type": "array"
					},
					"reminderTime": {
						"type": "number",
						"format": "double"
					},
					"deleted": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"siteId",
					"allowStartTime",
					"patrolMode",
					"site",
					"createdUserId",
					"createdUserName",
					"assignedUsers",
					"reminderTime",
					"deleted"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FindAndCountResponse_RouteResponse_": {
				"description": "Response with data array, count, limit and offset. Suited for Paging",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/RouteResponse"
						},
						"type": "array"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RouteDeleteResult": {
				"description": "response on route deletion",
				"properties": {
					"checkpointsDeleteCount": {
						"type": "number",
						"format": "double"
					},
					"schedulesDeleteCount": {
						"type": "number",
						"format": "double"
					},
					"routeDeleteCount": {
						"type": "number",
						"format": "double"
					},
					"routeName": {
						"type": "string"
					}
				},
				"required": [
					"checkpointsDeleteCount",
					"schedulesDeleteCount",
					"routeDeleteCount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RouteCheckpointResponse": {
				"description": "response of checkpoint apis",
				"properties": {
					"id": {
						"type": "string"
					},
					"setOrder": {
						"type": "number",
						"format": "double"
					},
					"layoutRow": {
						"type": "number",
						"format": "double"
					},
					"layoutCol": {
						"type": "number",
						"format": "double"
					},
					"cameras": {
						"items": {
							"$ref": "#/components/schemas/CameraResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"setOrder",
					"layoutRow",
					"layoutCol",
					"cameras"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DateOfWeek": {
				"description": "Sun, Mon, ..., Sat",
				"enum": [
					0,
					1,
					2,
					3,
					4,
					5,
					6
				],
				"type": "number"
			},
			"ExecuteTime": {
				"properties": {
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"endTime",
					"startTime"
				],
				"type": "object",
				"description": "start and end times"
			},
			"RouteScheduleResponse": {
				"description": "response of schedule apis",
				"properties": {
					"id": {
						"type": "string"
					},
					"startOccurrenceDate": {
						"type": "string",
						"format": "date-time"
					},
					"endOccurrenceDate": {
						"type": "string",
						"format": "date-time"
					},
					"executingDays": {
						"items": {
							"$ref": "#/components/schemas/DateOfWeek"
						},
						"type": "array"
					},
					"executingTime": {
						"items": {
							"$ref": "#/components/schemas/ExecuteTime"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"startOccurrenceDate",
					"endOccurrenceDate",
					"executingDays",
					"executingTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RouteDetailResponse": {
				"description": "response for route detail including schedules and checkpoints",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"siteId": {
						"type": "number",
						"format": "double"
					},
					"allowStartTime": {
						"type": "number",
						"format": "double"
					},
					"patrolMode": {
						"type": "string"
					},
					"site": {
						"$ref": "#/components/schemas/EntityOverallData"
					},
					"createdUserId": {
						"type": "string"
					},
					"createdUserName": {
						"type": "string"
					},
					"assignedUsers": {
						"items": {
							"$ref": "#/components/schemas/EntityOverallData"
						},
						"type": "array"
					},
					"reminderTime": {
						"type": "number",
						"format": "double"
					},
					"deleted": {
						"type": "boolean"
					},
					"routeCheckpoints": {
						"items": {
							"$ref": "#/components/schemas/RouteCheckpointResponse"
						},
						"type": "array"
					},
					"routeSchedules": {
						"items": {
							"$ref": "#/components/schemas/RouteScheduleResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"siteId",
					"allowStartTime",
					"patrolMode",
					"site",
					"createdUserId",
					"createdUserName",
					"assignedUsers",
					"reminderTime",
					"deleted"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RouteRequestBody": {
				"description": "request body for update and creation of route",
				"properties": {
					"name": {
						"type": "string"
					},
					"siteId": {
						"type": "number",
						"format": "double"
					},
					"allowStartTime": {
						"type": "number",
						"format": "double"
					},
					"patrolMode": {
						"type": "string"
					},
					"assignedUserIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"reminderTime": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"siteId",
					"allowStartTime",
					"patrolMode",
					"assignedUserIds",
					"reminderTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Entity": {
				"description": "Entity data is produced.",
				"properties": {
					"data": {}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EntityList": {
				"description": "List of entity data",
				"properties": {
					"entities": {
						"items": {
							"$ref": "#/components/schemas/Entity"
						},
						"type": "array"
					}
				},
				"required": [
					"entities"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RouteCheckpointRequest": {
				"description": "request body for update and creation of checkpoint",
				"properties": {
					"id": {
						"type": "string"
					},
					"setOrder": {
						"type": "number",
						"format": "double"
					},
					"layoutRow": {
						"type": "number",
						"format": "double"
					},
					"layoutCol": {
						"type": "number",
						"format": "double"
					},
					"cameraIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"setOrder",
					"layoutRow",
					"layoutCol",
					"cameraIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RouteScheduleRequest": {
				"description": "request body for update and creation of schedule",
				"properties": {
					"id": {
						"type": "string"
					},
					"startOccurrenceDate": {
						"type": "string",
						"format": "date-time"
					},
					"endOccurrenceDate": {
						"type": "string",
						"format": "date-time"
					},
					"executingDays": {
						"items": {
							"$ref": "#/components/schemas/DateOfWeek"
						},
						"type": "array"
					},
					"executingTime": {
						"items": {
							"$ref": "#/components/schemas/ExecuteTime"
						},
						"type": "array"
					}
				},
				"required": [
					"startOccurrenceDate",
					"endOccurrenceDate",
					"executingDays",
					"executingTime"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Authorization": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Bearer Token get from Login API (JWT)"
			}
		}
	},
	"info": {
		"title": "swagger-service",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "VPS"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/route/list": {
			"get": {
				"operationId": "ListRoutes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FindAndCountResponse_RouteResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/route/count": {
			"get": {
				"operationId": "CountRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CountResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "searchText",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/route/{id}": {
			"get": {
				"operationId": "GetRouteById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteRouteById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteDeleteResult"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteDetailResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/route": {
			"post": {
				"operationId": "CreateRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteDetailResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RouteRequestBody"
							}
						}
					}
				}
			}
		},
		"/route/{routeId}/checkpoint/list": {
			"get": {
				"operationId": "ListCheckpoints",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityList"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "routeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCheckpointByRouteId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteCheckpointResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "routeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RouteCheckpointRequest"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateCheckpointByRouteId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteCheckpointResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "routeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RouteCheckpointRequest"
							}
						}
					}
				}
			}
		},
		"/route/checkpoint/{id}": {
			"get": {
				"operationId": "GetCheckpointById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Entity"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/route/checkpoint/list": {
			"get": {
				"operationId": "DeleteCheckpoints",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "ids",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/route/{routeId}/schedule/list": {
			"get": {
				"operationId": "ListSchedules",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityList"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "routeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateScheduleByRouteId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteScheduleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "routeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RouteScheduleRequest"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateScheduleByRouteId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RouteScheduleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "routeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RouteScheduleRequest"
							}
						}
					}
				}
			}
		},
...
		"/route/schedule/list": {
			"get": {
				"operationId": "DeleteSchedules",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [
					{
						"Authorization": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "ids",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:9000/apis",
			"description": "Local environment"
		}
	]
}